✨(backend) create a file model
The file model will be the base model used by any resources uploadable
in Marsha.
