+++ src/backend/marsha/core/models/__init__.py
@@ -1,5 +1,6 @@
 """Make all models available from marsha.core.models."""
 # pylint: disable=wildcard-import,unused-wildcard-import
 from .account import *  # noqa isort:skip
+from .file import *  # noqa isort:skip
 from .playlist import *  # noqa isort:skip
 from .video import *  # noqa isort:skip
